拡張子やLoading処理を修正
- js -> jsx
- Loadingが生のテキストで出ていたのを修正
