+++ app/views/admin/courses/index.html.slim
@@ -12,4 +12,6 @@ header.page-header
 
 = render 'admin/admin_page_tabs'
 
-= react_component 'AdminCourses'
+.page-body
+  .container.is-lg
+    = react_component 'AdminCourses'
